refactor: UI 로직 분리 및 개선
- Game 관련 UI 로직을 views 폴더로 이동
- announce에서 render로 UI 함수명 변경
