+++ src/domain/game/utils.js
